# Add OPpTARGET_MY optimization to OP_UNDEF