--- a/t/perf/benchmarks
+++ b/t/perf/benchmarks
@@ -1350,6 +1350,22 @@
 
     # scalar assign, OP_SASSIGN
    
+    'expr::sassign::undef_lex' => {
+        setup   => 'my $x',
+        code    => '$x = undef',
+    },
+    'expr::sassign::undef_lex_direc' => {
+        setup   => 'my $x',
+        code    => 'undef $x',
+    },
+    'expr::sassign::undef_my_lex' => {
+        setup   => '',
+        code    => 'my $x = undef',
+    },
+    'expr::sassign::undef_my_lex_direc' => {
+        setup   => '',
+        code    => 'undef my $x',
+    },
 
     'expr::sassign::my_conststr' => {
         setup   => '',