… lib/B/Op_private.pm
… opcode.h