@richardleach 2022-08-22
Rebased following OP_PADSV_STORE merge. Awaiting some reviewers.