@haarg 2022-08-23
This isn't just an optree optimization. `undef $x` and `$x = undef` have different behavior. `$x = undef` leaves a PV allocated, but clears its flags. This means re-using the variable for a string of the same size will not allocate any new memory.

Without this patch:
```
perl -e'use Devel::Peek; my $f; Dump $f; $f = "x" x 40; Dump($f); $f = undef; Dump($f); undef $f; Dump($f);'
SV = NULL(0x0) at 0x1370226d0
  REFCNT = 1
  FLAGS = ()
SV = PV(0x13700b670) at 0x1370226d0
  REFCNT = 1
  FLAGS = (POK,pPOK)
  PV = 0x600000a28210 "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"\0
  CUR = 40
  LEN = 42
SV = PV(0x13700b670) at 0x1370226d0
  REFCNT = 1
  FLAGS = ()
  PV = 0x600000a28210 "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"\0
  CUR = 40
  LEN = 42
SV = PV(0x13700b670) at 0x1370226d0
  REFCNT = 1
  FLAGS = ()
  PV = 0
```
With this patch:
```
./perl -Ilib -e'use Devel::Peek; my $f; Dump $f; $f = "x" x 40; Dump($f); $f = undef; Dump($f); undef $f; Dump($f);'
SV = NULL(0x0) at 0x13880df80
  REFCNT = 1
  FLAGS = ()
SV = PV(0x13800b670) at 0x13880df80
  REFCNT = 1
  FLAGS = (POK,pPOK)
  PV = 0x600001172c70 "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"\0
  CUR = 40
  LEN = 42
SV = PV(0x13800b670) at 0x13880df80
  REFCNT = 1
  FLAGS = ()
  PV = 0
SV = PV(0x13800b670) at 0x13880df80
  REFCNT = 1
  FLAGS = ()
  PV = 0
```

This may be a reasonable change, but would at least need to be noted in the commit message.