@richardleach 2022-08-23
> This isn't just an optree optimization. `undef $x` and `$x = undef` have different behavior. `$x = undef` leaves a PV allocated, but clears its flags. This means re-using the variable for a string of the same size will not allocate any new memory.

Good point. I think the existing behaviour should be preserved. If there's a flag spare, it looks straightforward to implement.