@demerphq 2022-08-23
On Tue, 23 Aug 2022 at 15:25, demerphq ***@***.***> wrote:
>
> On Tue, 23 Aug 2022 at 13:52, Richard Leach ***@***.***> wrote:
...
> > Good point. I think the existing behaviour should be preserved. If there's a flag spare, it looks straightforward to implement.
>
> I vote that the existing behavior be preserved, it can have some
> interesting performance implications.

Sorry i phrased that poorly. I understand you just said you would do
that, i just meant to add that there is good reason to do so.
Appreciate your efforts on this stuff.

cheers,
Yves

-- 
perl -Mre=debug -e "/just|another|perl|hacker/"