@richardleach 2022-08-23
> On Tue, 23 Aug 2022 at 15:25, demerphq ***@***.***> wrote:
> I vote that the existing behavior be preserved, it can have some interesting performance implications.

That's okay, I understood it as you meant it.

The current branch should now preserve the pre-existing behaviour w.r.t. PVs:
```
perl5.37.4 -e'use Devel::Peek; my $f; Dump $f; $f = "x" x 40; Dump($f); $f = undef; Dump($f); undef $f; Dump($f);'
SV = NULL(0x0) at 0x5636bb8f1720
  REFCNT = 1
  FLAGS = ()
SV = PV(0x5636bb8cdff0) at 0x5636bb8f1720
  REFCNT = 1
  FLAGS = (POK,pPOK)
  PV = 0x5636bb8f8ed0 "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"\0
  CUR = 40
  LEN = 42
SV = PV(0x5636bb8cdff0) at 0x5636bb8f1720
  REFCNT = 1
  FLAGS = ()
  PV = 0x5636bb8f8ed0 "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"\0
  CUR = 40
  LEN = 42
SV = PV(0x5636bb8cdff0) at 0x5636bb8f1720
  REFCNT = 1
  FLAGS = ()
  PV = 0
```

I've added the tests you suggested to _t/perf/opcount.t_, will just tidy them slightly so they don't break unnecessarily in the next opcode update.