@richardleach 2022-08-23
Ok. tests tidied. I think all review comments to date have been addressed.